% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_fonts.R
\name{adj_fonts}
\alias{adj_fonts}
\title{Adjust font sizes of ggplot object}
\usage{
adj_fonts(big = 12, small = 10)
}
\arguments{
\item{big}{int, default=12; axis and legend title font size}

\item{small}{int, default=10; axis markings and legend item font size}
}
\value{
ggplot theme object to adjust fonts
}
\description{
Adjust font sizes of ggplot object
}
