% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric_dt.R
\name{as.numeric.dt}
\alias{as.numeric.dt}
\title{Convert data table with 1 row to named vector}
\usage{
\method{as.numeric}{dt}(x)
}
\arguments{
\item{x}{data.table with single row}
}
\value{
named numeric vector
}
\description{
Convert data table with 1 row to named vector
}
