% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric_factor.R
\name{as.numeric.factor}
\alias{as.numeric.factor}
\title{Convert factor to numeric}
\usage{
\method{as.numeric}{factor}(x)
}
\arguments{
\item{x}{factor vector}
}
\value{
numeric vector
}
\description{
Convert factor to numeric
}
