% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_legend.R
\name{g_legend}
\alias{g_legend}
\title{Return legend from ggplot object}
\usage{
g_legend(gplot)
}
\arguments{
\item{gplot}{ggplot object from which to get legend}
}
\value{
legend from gplot
}
\description{
Return legend from ggplot object
}
