% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_box.R
\name{input_box}
\alias{input_box}
\title{Open an input box}
\usage{
input_box(message = "")
}
\arguments{
\item{message}{message to display to user}
}
\value{
use input as string
}
\description{
Open an input box
}
