% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_legend.R
\name{no_legend}
\alias{no_legend}
\title{Return ggplot object without legend}
\usage{
no_legend(gplot)
}
\arguments{
\item{gplot}{ggplot object from which to remove legend}
}
\value{
ggplot without legend
}
\description{
Return ggplot object without legend
}
