% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_restrict.R
\name{restrict_statistics}
\alias{restrict_statistics}
\alias{meanr}
\alias{medianr}
\alias{mean_ser}
\title{Restrict statistics by number of observations}
\usage{
meanr(x, n = 5, ...)

medianr(x, n = 5, ...)

mean_ser(x, n = 5, ...)
}
\arguments{
\item{x}{numeric vector; observations}

\item{n}{integer; number of observations to consider valid sample}

\item{...}{arguments passed to main function (e.g. NA.rm)}
}
\value{
result of call
}
\description{
Restrict statistics by number of observations
}
