% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_abw.R
\name{theme_abw}
\alias{theme_abw}
\title{Custom ggplot theme, a modified theme_bw}
\usage{
theme_abw(
  big_font = 14,
  small_font = 12,
  grid = F,
  x_adjust = -0.5,
  y_adjust = 2,
  x_angle = 0,
  legend = T
)
}
\arguments{
\item{big_font}{int, default=12; axis and legend title font size}

\item{small_font}{int, default=10; axis markings and legend item font size}

\item{grid}{logical, default=FALSE; display gridlines?}

\item{x_adjust}{numeric, default=-0.05; placement of x axis title}

\item{y_adjust}{numeric, default=1; placement of y axis title}

\item{x_angle}{numeric, default=0; angle of x axis markings}

\item{legend}{logical, default=TRUE; display legend?}
}
\value{
custom ggplot theme object
}
\description{
Custom ggplot theme, a modified theme_bw
}
